refactor(sandbox-fc): isolate snapshot process lifecycle
