+++ crates/sandbox-fc/src/snapshot/attempt/tests.rs
@@ -5,6 +5,7 @@ use std::time::Duration;
 use nbd_cow::KeptCow;
 use nbd_cow::pool::DevicePoolHandle;
 
+use crate::api::ApiError;
 use crate::paths::{SandboxPaths, SnapshotOutputPaths, SockPaths};
 use crate::snapshot::cow::{snapshot_attempt_cow_file, snapshot_attempt_workspace_image_file};
 use crate::snapshot::publish::SnapshotPublishAttempt;
@@ -440,6 +441,22 @@ fn long_running_child_for_test() -> tokio::process::Child {
         .expect("spawn long-running child")
 }
 
+async fn exited_child_for_test() -> tokio::process::Child {
+    let mut child = tokio::process::Command::new("sh")
+        .arg("-c")
+        .arg("exit 1")
+        .stdin(std::process::Stdio::null())
+        .stdout(std::process::Stdio::null())
+        .stderr(std::process::Stdio::null())
+        .process_group(0)
+        .kill_on_drop(true)
+        .spawn()
+        .expect("spawn exiting child");
+    let status = child.wait().await.expect("wait for exiting child");
+    assert!(!status.success(), "test child should exit unsuccessfully");
+    child
+}
+
 #[tokio::test]
 async fn snapshot_attempt_drop_handoff_releases_netns_without_unlocked_sock_cleanup() {
     let dir = tempfile::tempdir().expect("tempdir");
@@ -486,6 +503,45 @@ async fn snapshot_attempt_drop_handoff_kills_child_before_netns_release() {
         tokio::fs::try_exists(&sock_dir).await.unwrap(),
         "detached cleanup must not remove the stable snapshot socket directory without the outer snapshot lock"
     );
+}
+
+#[tokio::test(start_paused = true)]
+async fn snapshot_attempt_cancelled_finish_hands_complete_process_to_finalizer() {
+    let dir = tempfile::tempdir().expect("tempdir");
+    let (mut attempt, _sock_dir) = snapshot_attempt_for_test(&dir);
+
+    attempt.track_child_for_test(exited_child_for_test().await);
+    attempt.track_stderr_handle_for_test(tokio::spawn(std::future::pending::<()>()));
+
+    let finish = tokio::time::timeout(
+        Duration::from_millis(1),
+        attempt.finish_runtime_after_workflow(Err(SnapshotError::Api(ApiError::Other(
+            "timeout".into(),
+        )))),
+    )
+    .await;
+    assert!(
+        finish.is_err(),
+        "explicit finish should be cancelled during stderr drain"
+    );
+
+    let presence = attempt.cleanup_resources.presence();
+    assert!(
+        presence.has_child,
+        "cancelled finish must retain the child owner"
+    );
+    assert!(
+        presence.has_stderr_forwarder,
+        "cancelled finish must retain the stderr forwarder"
+    );
+
+    let (tx, rx) = tokio::sync::oneshot::channel();
+    attempt.notify_cleanup_complete_for_test(tx);
+    drop(attempt);
+    let report = wait_for_snapshot_cleanup(rx).await;
+
+    assert!(report.child_reaped);
+    assert!(report.stderr_forwarder_finished);
 }
 
 #[tokio::test]
